fix(webapp): move createTrayFetch out of scoops/ entirely
Round 1 finding: caching isChromeExtensionRealm() in base/api-endpoint.ts
and reading the cache from scoops/tray-leader.ts was the same probe under
a new name, still in scoops/ business logic. A slice-D lint gate keyed on
probe names would have been bypassed by exactly that move.

createTrayFetch (and its TrayProxyFetchError) moves to shell/tray-fetch.ts,
a sibling of proxied-fetch.ts rather than folded into it (keeps the eager
graph delta at zero — folding in would pull proxied-fetch.ts's larger
SecureFetch surface into anything that only needs the tray transport).
Deciding raw fetch vs /api/fetch-proxy is a transport-layer decision;
scoops/tray-leader.ts now has no realm or topology read anywhere in the
file. LeaderTrayManager still does `options.fetchImpl ?? createTrayFetch()`,
importing the factory from shell/ (a downward import); the module
re-exports both names under their established address so no other caller
changes.

redirect-uri.ts's callers resolving resolveFloatTopology() at their own
call site is correct as-is and unchanged — both callers are shell/, which
owns topology; that was never a relocation that needed fixing.

Also:
- base/api-endpoint.ts's getChromeExtensionRealm/setChromeExtensionRealm
  doc comments now say explicitly that this is still a probe (a deduped
  read, not a relocated decision), and that scoops/, tools/ and kernel/
  (except kernel/host.ts) must never call it — noted as a name slice D's
  gate will ban, alongside isExtensionRealm/hasLocalNodeServer/
  resolveFloatTopology.
- The guard test scans tray-leader.ts's WHOLE source (not just its import
  statements) for all six probe names, so a call-site alias or re-derived
  probe would still fail it, not just a literal import. Verified against a
  regression by reintroducing the old branch with Edit, confirming the
  test fails, then reverting.
- capability/index.ts and docs/work-unit.md (phase 6c) no longer claim
  "network is done" as if the call sites moved onto the broker: none of
  the three is a privileged operation behind an allowlist, so none became
  network.crossOriginFetch. What actually happened is scoops/ lost its
  only realm/topology read.

First-load: +0.0 kB on both graphs (measured), confirming the sibling
module didn't pull tray-leader.ts's other imports onto the eager graph.

Part of #2276
