+++ packages/webapp/src/work-unit/capability/index.ts
@@ -4,28 +4,40 @@
  * Slice A shipped the protocol and one page stub. Slice B replaced the stubs
  * with real adapters keyed by float topology (`node-rest`, `extension-direct`,
  * `extension-delegate`, `connect`), composed once in `kernel/host.ts`. Slice C
- * migrates the remaining privileged call sites off direct float probes, one
- * PR per domain — network is done.
+ * removes float probes from `scoops/` / `tools/` / `kernel/` business logic
+ * (review-patterns category 10), one domain per PR — network is done.
  *
- * `network` (#2276 slice C, done): `scoops/tray-leader.ts` (`createTrayFetch`),
- * `shell/proxied-fetch.ts` (`createProxiedFetch`) and `shell/mcp/redirect-uri.ts`
- * were NOT rewired onto `broker.network.crossOriginFetch` — they are
- * `SecureFetch`-shaped transport factories called from 18+ shell sites with
- * no broker in scope, not privileged operations behind an allowlist, so
- * threading a broker instance through all of them would have been a
- * different, much larger refactor. Instead each call site takes the ANSWER
- * by injection rather than probing: `tray-leader.ts` and `proxied-fetch.ts`
- * read `getChromeExtensionRealm()` (`base/api-endpoint.ts`'s lazily-cached,
- * per-realm fact — resolved once, not re-probed per call, mirroring the
- * existing `extensionDelegateId` / `localApiBaseUrl` idiom in that same
- * module); `redirect-uri.ts`'s `resolveMcpRedirectUri` takes `topology` as a
- * parameter, resolved by its two callers at the point they actually need a
- * redirect URI. See `docs/work-unit.md` phase 6c.
+ * `network` (#2276 slice C, done): none of `scoops/tray-leader.ts`
+ * (`createTrayFetch`), `shell/proxied-fetch.ts` (`createProxiedFetch`) or
+ * `shell/mcp/redirect-uri.ts` (`resolveMcpRedirectUri`) is now `broker.network
+ * .crossOriginFetch` — none of the three is a privileged operation behind an
+ * allowlist, they are `SecureFetch`-shaped transport factories called from
+ * 18+ `shell/` sites with no broker in scope, so `network.crossOriginFetch`
+ * stays the broker op for a caller that DOES hold a broker (none migrated
+ * yet, and none need to for this reason). What actually happened: `scoops/
+ * tray-leader.ts` lost its only realm/topology read, period — `createTrayFetch`
+ * (with its `TrayProxyFetchError`) moved to `shell/tray-fetch.ts`, a sibling
+ * of `proxied-fetch.ts`, because deciding "raw fetch or `/api/fetch-proxy`"
+ * is a transport-layer decision, not `scoops/` business logic; caching the
+ * probe in place would have been the same probe under a new name, in the
+ * same wrong layer. `shell/proxied-fetch.ts` and `shell/tray-fetch.ts` read
+ * `getChromeExtensionRealm()` — `base/api-endpoint.ts`'s lazily-cached,
+ * per-realm fact, mirroring its `extensionDelegateId` / `localApiBaseUrl`
+ * idiom — because `shell/` is where topology is OWNED (see
+ * `shell/float-topology.ts`'s header); that read is fine and expected there,
+ * it is still a probe, and `scoops/` / `tools/` / `kernel/` (except `kernel/
+ * host.ts`) must never call it either. `redirect-uri.ts`'s
+ * `resolveMcpRedirectUri` takes `topology` as a parameter, resolved by its
+ * two callers (also `shell/`) at the point they actually need a redirect
+ * URI — the same ownership rule, not a relocation to fix. See
+ * `docs/work-unit.md` phase 6c.
  *
- * TODO(#2276) slice C, remaining domains — privileged call sites that still
- * branch on float / topology (`isChromeExtensionRealm` / `isExtensionRealm` /
- * `hasLocalNodeServer` / `resolveFloatTopology`) rather than this broker.
- * One PR per domain, smallest first:
+ * TODO(#2276) slice C, remaining domains — privileged call sites in `scoops/`
+ * / `tools/` / `kernel/` that still branch on float / topology
+ * (`isChromeExtensionRealm` / `isExtensionRealm` / `hasLocalNodeServer` /
+ * `resolveFloatTopology` / `getChromeExtensionRealm` /
+ * `setChromeExtensionRealm`) rather than asking an injected `CapabilityBroker`
+ * or taking a composition-time answer. One PR per domain, smallest first:
  *
  * secrets
  *   - scoops/scoop-context/shell-and-skills.ts `fetchSecretEnvVars`
